requirements.txt
Flask
python-dotenv
google-generativeai
langchain
pinecone-client
sentence-transformers
